Move `NtWriteFile` miri shim to `foreign_items`
